Update examples.md
Added an example of blinking an LED with a passive millis counting delay after the "Hello World" example.  Unfortunately, this does not solve the problem of a long loop breaking cloud connectivity, and might actually make things worse if v.02 is only looking for long delay() for background processing.  Checking in for now, but will probably abandon this route.